Fetch and parse turtle output from Worldcat & CrossRef
This was a frustrating one.

First off, I create the turtle URLs by hand because of two different
issues. The simpler one is that WorldCat hasn't yet enabled CORS for
their Turtle resources. The more complex one is that web browsers act
differently when they receive a 303 SEE OTHER response. Seeing a CORS
request, Firefox will preserve the `Accept` header. Chrome does not.
There's no way to change the redirect request either- the redirection is
done transparently by the browser.

Chrome will be fixed in version 36 [1], and WorldCat will probably come
around on the CORS front soon. Blah

Otherwise, this is all pretty neat. The turtle triples are parsed and
stored using the N3 library, and then we pluck out the data that we need
(title, date, authorship) and set it on the source model. Some of the
code calls for a refactor or splitting out of components, but that can
come later.

[1] https://code.google.com/p/chromium/issues/detail?id=162183