Make persistence drastically dumber and simpler
Instead of denormalizing things and worrying about fetching
relationships, just store the whole structure of the dataset in IDB and
keep it in memory when on the site.

Sorry, I finally had enough of trying to deal with relationships in
IndexedDB. Hopefully someday they'll get better with indexing so that
it's possible to make functional/arbitrary indexes or something.