Improve syncing to only deal with remote changes
When syncing against a remote data source, I've added a function that
allows you to only deal with changes that originate from the remote
server.

For example, if I were to add a period to my own collection, and then
synced to another data source, the diff would tell me that I should
delete my new collection in order to get to the state of the remote
source. However, I'm not interested in this change for the purposes of
syncing, since it did not come from the remote server.

This approach will allow us to only get changes initiated locally as
well, for the purpose of, say, generating a patch to a remote server.