Change how patches are generated
This commit changes the patch.makePatch function that generates a
JSON-Patch that transforms one object to another. We had been using
fast-json-patch's `compare` function to do this, which implemented an
algorithm which would 1) traverse the tree 2) delete/add leaf nodes only
present in the source/destination 3) replace leaf nodes in both the
source and destination with the value in the destination.

That is a smart-enough algorithm for most purposes, but we need
something different, since we store and search for JSON-Patch records in
order to determine whether a certain change in syncing/patching was
something that was already done in the client (see 481ff7).

Instead of fast-json-patch's straightforward leaf-based algorithm, we
now generate patches slightly differently:
  1) Generate the previous json-fast-patch.compare patch
  2) Iterate through every patch, performing the following checks
     a) If the patch is an add/remove op on a period or period
        collection, keep it
     b) Check if the attribute being modified is a 'top-level' attribute
        of a period or collection (i.e. `label`, but not
        `localizedLabel.en`, which is a second level attribute)
     c) If the patch is changing a top-level attribute with an add or
        remove operation, keep it
     d) For any other patch (replace for any op, or add/remove for
        non-top-level attributes), create two new patches: a remove for
        the relevant top-level attribute, followed by an add for the
        entire value of the top-level attribute. Keep a reference that
        this top-level attribute has already been replaced in its
        entirety, and skip anymore patches that affect it.

By carrying out those steps, we will be able to have an idea of what the
value of a period or collection's attributes are at any client revision
without having to reconstruct their histories.

I've also updated the PatchDiffCollection to allow filtering based on
patches generated by the new algorithm.