Update period edit view to use Immutable.js
This was a biggie. I'm not going to go through everything- the gist of
it was to get away from two-way data binding using backbone.stickit and
instead *not* track state. The current data of the form can be accessed
with .getData(). This is not ideal, and I wouldn't do it this way over
again, but the benefits of Immutable trump the annoyance of doing
explicit data binding/retrieving.