Don't choke on ill-formed periods
This makes it so that the application doesn't break when fields like
start, stop, and originalLabel are not present. Theoretically, they
should *always* be around, but this gives us more flexibility. Also,
React likes to catastrophically explode when things go wrong during
render(), so error recovery is a work in progress in this application.