First stab at having options that can be updated within a resource
I don't really care for storing JSON in the url params, but we need to
be able to store arrays and nested objects. It's either this, use a
technique like the `qs` library (which transforms 'a[]=1&a[]=2' into
`a=[1,2]`), or abandon the goal of storing application state in URLs.
The latter is out, and I don't like `qs`, so this is it for the time
being.