Update querystring logic so that it's up to date with Node 5+
The version bundled with broserify was out of date. It did not
recognize the `encodeURIComponent`/`decodeURIComponent` options for
stringify/parse, respectively. I was using those to see if we could
make the URLs slightly less ugly. That's not a real priority, so I
won't mess with it anymore. But if we wanted to in the future, it's
possible now.