Lots of makefile changes
Building a zip file was pretty broken before because of an error in the
sed oneliner I wrote in the `zip` target. I fixed that, added images to
the zip distribution, and changed the name of the folder containing
images from `assets` to `images`.

Also, lots of little tweaks to the makefile. (Most notably, using the
locally installed version of lerna instead of relying on it being in
`$PATH`)