Fix viewing Web backend history + allow viewing individual Web patches
We were using jsonld.frame to generate the JSON for the backend
history, but it was waaaay to slow for the current dataset (which is
not even very big). It seems like the framing function grew
exponentially for each new entry in the changelog. Instead, we now just
query with N3 directly to generate a representation of a patch. That
same query logic is used to generate individual diffs as well.