Fix rendering problems in TextValue and Source
I knew there was something fishy about how note fields were being
rendered. For every link found in the text, the entire value of the
field up to that point was appended to the final value, rather than the
value from the end of the previously found link. This fixes that.

Also....

* Preserve white space

* Make Web page field on Source component a LinkValue