Add file backends back in
Involved a lot of shuffling around due to the fact that
BackendStorage.StaticFile cannot be constructed with just a string.

I could have made it the same as BackendStorage.IndexedDB, but it would
have required adding a third argument to the CreateBackend action to
allow the file to be submitted at the same time as the newly created
StaticFile backend. The IndexedDB backend storage does not have the same
problem because it always starts as an empty dataset.