Add the notion of "colorsets"
These are different sets of background, foreground, and border colors
that can be applied to different elements. A couple are specific to
elements (table, tableSecondary, menu), others are general categories
(primary, secondary). Good enough!

Not all elements are colored with colorsets, but we can deal with that
along the way. Again, good enough!