Move breadcrumb logic to the resource definition level
* Resources can define their breadcrumbs in a `getBreadcrumb` attribute
of their own definition

* The breadcrumb component now takes a list of breadcrumbs, with three
attributes: label, route, truncate. If a route is present, the crumb is
turned into a link. If truncate is true, the crumb will be truncated if
it's too long. If it just has a label, it'll be text.

* All breadcrumb rendering was taking out of resource components

* I was going to add the breadcrumb rendering to the withNavigation HOC
from 4e4d120, but breadcrumb rendering needed all of the props that
would be passed to the final resource component, so I added it at the
very end of the chain.