Don't load data source history for browse authorities page
It used to be "my authorities"-- a page to browse the authorities that
one had made themselves. In order to see what authorities someone had
made, we had to go through all the patches created in a local backend.
Since it's now "Browse authorities", nothing needs to be done-- the
component will already be passed the dataset, and can get the
authorities directly from there.