Update the way we filter out patches between datasets
The previous way was really strict about what it would let through to
send to remote datasets, and didn't allow deleting items. This new way
is more explicit and only filters out a minimal amount of things. It's
also much better documented. See the notes in `makeFilteredPatch` for
more details