Add a `validatePatch` function
This function takes a raw dataset and a patch, creates an intermediary
DatasetProxy from the result of applying the patch to the dataset, and
then returns a new patch that is the result of going from the original
dataset to the validated dataset created by the DatasetProxy.

See dae853c for more details. Essentially, this is to prevent submitting
patches that result in an invalid dataset-- such as including related
periods that will not exist after applying the patch.