Refactor how authority layout deals with selected periods
1. Use org-shell's API to deal with updating dynamic options. This means
especially, using org-shell's `opts` to store whether a period has been
selected, rather than its `params`, which are meant to be static.

2. To do this, we have to pass org-shell's `opts` and `updateOpts`,
which (confusingly, sorry) are different than the layout renderer's
`opts` and `updateOpts`, which only allow getting/setting options for *a
single layout block*. Since the selected period affects multiple blocks,
and we also want to persist it in the URL, this means we need to use the
org-shell opts/updateOpts, which affect the whole resource.

3. Don't have the windowed period list determine whether the selected
period is present in the filtered dataset. (The selected period wouldn't be
present in the case that someone selected a period, and applied some
more filters, filtering out the selected period). Because the selected
period is persisted in the URL state, it is always passed down to
children of the layout. Rather than having the windowed period list
(which is the component at the end of the pipeline, so knows whether a
selected period is in the filtered data) tell the authority layout
renderer whether a selected period is in the filtered data, components
which do something with selected periods (the timeline and map) just
check whether the selected period is in the data they received.

(That was a very long third point that I could edit for brevity and
clarity, but I think you get what I am saying well enough)