Make sure the date parser is compiled before building project
As a consequence of f6b8a9f, our submodules are no longer have `npm
install` run on them when they are smushed together-- they are just
regular files. The date-parser was the only module that required
anything extra: running pegjs to compile the grammar into a javascript
file.

This commit makes sure that happens via the makefile. If we ever want to
update the date parser and publish it, we can still do so inside the
module. But this prevents have to keep track of another
node_modules/package-lock.json.

It might make sense to move periodo-date-parser to an external module,
also!